extract closeIcon to be standalone and reusable component
