+++ src/components/dismiss-button/dismiss-button-classic.style.js
@@ -1,6 +1,6 @@
 import { css } from 'styled-components';
-import { THEMES } from '../../../style/themes';
-import classicThemeColors from '../message-classic-theme-colors';
+import { THEMES } from '../../style/themes';
+import classicThemeColors from '../message/message-classic-theme-colors';
 
 export default ({ theme, messageType }) => theme.name === THEMES.classic && css`
   background-color: ${classicThemeColors[messageType].backgroundColor};
