+++ src/components/dismiss-button/index.js
@@ -1,0 +1,1 @@
+export { default } from './dismiss-button.component';
